refactor: split PatientsPage and MedicationsPage into modular components
